fix: refactor imports for better module organization in paper components
