+++ packages/joint-react/src/components/paper/render-element/paper-element-item.tsx
@@ -8,7 +8,8 @@ import {
 } from 'react';
 import { createPortal } from 'react-dom';
 import typedMemo from '../../../utils/typed-react';
-import { useGraphStore, usePaper } from '../../../hooks';
+import { useGraphStore } from '../../../hooks/use-graph-store';
+import { usePaper } from '../../../hooks/use-paper';
 import { useCellId } from '../../../hooks/use-cell-id';
 import { ELEMENT_MODEL_TYPE } from '../../../mvc/element-model';
 import type { CellId, ElementRecord } from '../../../types/cell.types';
